+++ src/main/java/chilling/encore/global/config/security/util/SecurityUtils.java
@@ -1,15 +1,12 @@
 package chilling.encore.global.config.security.util;
 
 import chilling.encore.domain.User;
-import chilling.encore.global.config.jwt.PrincipalDetails;
-import chilling.encore.global.exception.ApplicationException;
+import chilling.encore.global.config.security.jwt.PrincipalDetails;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Objects;
 import java.util.Optional;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
