+++ src/main/java/chilling/encore/service/UserService.java
@@ -7,8 +7,8 @@ import chilling.encore.dto.AlaramDto.AlarmResponse;
 import chilling.encore.dto.UserDto.UserLoginRequest;
 import chilling.encore.dto.UserDto.UserLoginResponse;
 import chilling.encore.dto.UserDto.UserSignUpRequest;
-import chilling.encore.global.config.jwt.JwtTokenProvider;
-import chilling.encore.global.config.jwt.TokenInfoResponse;
+import chilling.encore.global.config.security.jwt.JwtTokenProvider;
+import chilling.encore.global.config.security.jwt.TokenInfoResponse;
 import chilling.encore.global.config.redis.RedisRepository;
 import chilling.encore.global.config.security.util.SecurityUtils;
 import chilling.encore.repository.springDataJpa.CenterRepository;
